refactor: addClass 컴포넌트 page index 구조로 변경
